Module ferrisetw::native::tdh_types

Basic TDH types

The tdh_type module provides an abstraction over the basic TDH types, this module act as a helper for the parser to determine which IN and OUT type are expected from a property within an event

This is a bit extra but is basically a redefinition of the In an Out TDH types following the rust naming convention, it can also come in handy when implementing the TryParse trait for a type to determine how to handle a Property based on this values

Structs

Property
Attributes of a property

PropertyFlags
Represents the Property flags

Enums

TdhInType
Represent a TDH_IN_TYPE

TdhOutType
Represent a TDH_OUT_TYPE