Struct ferrisetw::parser::Parser

pub struct Parser<'a> { /* fields omitted */ }

Represents a Parser

This structure holds the necessary data to parse the ETW event and retrieve the data from the event

Implementations

impl<'a> Parser<'a>

pub fn create(schema: &'a Schema) -> Self

Use the create function to create an instance of a Parser

Arguments

schema - The Schema from the ETW Event we want to parse

Example

let my_callback = |record: EventRecord, schema_locator: &mut SchemaLocator| {
    let schema = schema_locator.event_schema(record)?;
    let parser = Parse::create(&schema);
};