Trait ferrisetw::parser::TryParse

pub trait TryParse<T> {
    fn try_parse(&mut self, name: &str) -> Result<T, ParserError>;
}

Trait to try and parse a type

This trait has to be implemented in order to be able to parse a type we want to retrieve from within an Event. On success the parsed value will be returned within a Result, on error an Err should be returned accordingly

An implementation for most of the Primitive Types is created by using a Macro, any other needed type requires this trait to be implemented

Required methods

fn try_parse(&mut self, name: &str) -> Result<T, ParserError>

Implement the try_parse function to provide a way to Parse T from an ETW event or return an Error in case the type T can't be parsed

Arguments

name - Name of the property to be found in the Schema

Loading content...