Module ferrisetw::provider::kernel_providers

Kernel Providers module

Provides an easy way to create a Kernel Provider. Multiple providers are pre-created statically with their appropriate GUID and flags Credits: KrabsETW::kernel_providers

Modules

kernel_flags
List of Kernel Providers flags

kernel_guids
List of Kernel Providers GUIDs

Structs

ALPC_PROVIDER
Represents the ALPC Kernel Provider

CONTEXT_SWITCH_PROVIDER
Represents the Context Swtich Kernel Provider

DEBUG_PRINT_PROVIDER
Represents the Dbg Pring Kernel Provider

DISK_FILE_IO_PROVIDER
Represents the Disk File IO Kernel Provider

DISK_IO_INIT_PROVIDER
Represents the Disk IO Init Kernel Provider

DISK_IO_PROVIDER
Represents the Disk IO Kernel Provider

DPC_PROVIDER
Represents the DPC Kernel Provider

DRIVER_PROVIDER
Represents the Driver Kernel Provider

FILE_INIT_IO_PROVIDER
Represents the File Init IO Kernel Provider

FILE_IO_PROVIDER
Represents the File IO Kernel Provider

IMAGE_LOAD_PROVIDER
Represents the Image Load Kernel Provider

INTERRUPT_PROVIDER
Represents the Interrupt Kernel Provider

KernelProvider
Represents a Kernel Provider structure which can be used to create a Kernel Provider

MEMORY_HARD_FAULT_PROVIDER
Represents the Memory Hard Fault Kernel Provider

MEMORY_PAGE_FAULT_PROVIDER
Represents the Memory Page Fault Kernel Provider

PROCESS_COUNTER_PROVIDER
Represents the Process Counter Kernel Provider

PROCESS_PROVIDER
Represents the Process Kernel Provider

PROFILE_PROVIDER
Represents the Profile Kernel Provider

REGISTRY_PROVIDER
Represents the Registry Kernel Provider

SPLIT_IO_PROVIDER
Represents the Split IO Kernel Provider

SYSTEM_CALL_PROVIDER
Represents the SystemCall Kernel Provider

TCP_IP_PROVIDER
Represents the TCP-IP Kernel Provider

THREAD_DISPATCHER_PROVIDER
Represents the Thread Dispatcher Kernel Provider

THREAD_PROVIDER
Represents the Thread Kernel Provider

VAMAP_PROVIDER
Represents the VA Map Kernel Provider

VIRTUAL_ALLOC_PROVIDER
Represents the VirtualAlloc Kernel Provider