Trait ferrisetw::trace::TraceTrait

pub trait TraceTrait: TraceBaseTrait {
    fn named(self, name: String) -> Self;

    fn enable_provider(&self) { ... }

    fn augmented_file_mode() -> u32 { ... }

    fn enable_flags(_providers: &RwLock<Vec<Provider>>) -> u32 { ... }

    fn trace_guid() -> Guid { ... }
}

Specific trait for a Trace

This trait define the specific methods that differentiate from a Kernel to a User Trace

Required methods

fn named(self, name: String) -> Self

Use the named function to set the trace name

Arguments

name - Trace name to set

Remarks

If this function is not called during the process of building the trace a random name will be generated

Example

let my_trace = UserTrace::new().named("TestTrace".to_string());

Loading content...
Provided methods

fn enable_provider(&self)

fn augmented_file_mode() -> u32

fn enable_flags(_providers: &RwLock<Vec<Provider>>) -> u32

fn trace_guid() -> Guid

Loading content...